rocksdbccl: split the enterprise c++ code into the ccl tree
