+++ pkg/storage/engine/rocksdb.go
@@ -22,7 +22,6 @@ package engine
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"math"
 	"os"
 	"path/filepath"
@@ -318,6 +317,11 @@ func newMemRocksDB(attrs roachpb.Attributes, cache RocksDBCache, maxSize int64) 
 // String formatter.
 func (r *RocksDB) String() string {
 	return fmt.Sprintf("%s=%s", r.attrs.Attrs, r.dir)
+}
+
+// RawEngine returns the wrapped cgo DBEngine.
+func (r *RocksDB) RawEngine() *C.DBEngine {
+	return r.rdb
 }
 
 func (r *RocksDB) open() error {
@@ -1502,112 +1506,4 @@ func dbIterate(
 	}
 	// Check for any errors during iteration.
 	return it.Error()
-}
-
-// RocksDBSstFileReader allows iteration over a number of non-overlapping
-// sstables exported by `RocksDBSstFileWriter`.
-type RocksDBSstFileReader struct {
-	// TODO(dan): This currently works by creating a RocksDB instance in a
-	// temporary directory that's cleaned up on `Close`. It doesn't appear that
-	// we can use an in-memory RocksDB with this, because AddFile doesn't then
-	// work with files on disk. This should also work with overlapping files.
-
-	dir     string
-	rocksDB *RocksDB
-}
-
-// MakeRocksDBSstFileReader creates a RocksDBSstFileReader that uses a scratch
-// directory which is cleaned up by `Close`.
-func MakeRocksDBSstFileReader() (RocksDBSstFileReader, error) {
-	dir, err := ioutil.TempDir("", "RocksDBSstFileReader")
-	if err != nil {
-		return RocksDBSstFileReader{}, err
-	}
-
-	// TODO(dan): I pulled all these magic numbers out of nowhere. Make them
-	// less magic.
-	cache := NewRocksDBCache(1 << 20)
-	rocksDB, err := NewRocksDB(
-		roachpb.Attributes{}, dir, cache, 512<<20, DefaultMaxOpenFiles)
-	if err != nil {
-		return RocksDBSstFileReader{}, err
-	}
-	return RocksDBSstFileReader{dir, rocksDB}, nil
-}
-
-// AddFile links the file at the given path into a database. See the RocksDB
-// documentation on `AddFile` for the various restrictions on what can be added.
-func (fr *RocksDBSstFileReader) AddFile(path string) error {
-	if fr.rocksDB == nil {
-		return errors.New("cannot call AddFile on a closed reader")
-	}
-	return statusToError(C.DBEngineAddFile(fr.rocksDB.rdb, goToCSlice([]byte(path))))
-}
-
-// Iterate iterates over the keys between start inclusive and end
-// exclusive, invoking f() on each key/value pair.
-func (fr *RocksDBSstFileReader) Iterate(
-	start, end MVCCKey, f func(MVCCKeyValue) (bool, error),
-) error {
-	if fr.rocksDB == nil {
-		return errors.New("cannot call Iterate on a closed reader")
-	}
-	return fr.rocksDB.Iterate(start, end, f)
-}
-
-// Close finishes the reader.
-func (fr *RocksDBSstFileReader) Close() {
-	if fr.rocksDB == nil {
-		return
-	}
-	fr.rocksDB.Close()
-	fr.rocksDB = nil
-	if err := os.RemoveAll(fr.dir); err != nil {
-		log.Warningf(context.TODO(), "error removing temp rocksdb directory %q: %s", fr.dir, err)
-	}
-}
-
-// RocksDBSstFileWriter creates a file suitable for importing with
-// RocksDBSstFileReader.
-type RocksDBSstFileWriter struct {
-	fw *C.DBSstFileWriter
-	// DataSize tracks the total key and value bytes added so far.
-	DataSize int64
-}
-
-// MakeRocksDBSstFileWriter creates a new RocksDBSstFileWriter with the default
-// configuration.
-func MakeRocksDBSstFileWriter() RocksDBSstFileWriter {
-	return RocksDBSstFileWriter{C.DBSstFileWriterNew(), 0}
-}
-
-// Open creates a file at the given path for output of an sstable.
-func (fw *RocksDBSstFileWriter) Open(path string) error {
-	if fw == nil {
-		return errors.New("cannot call Open on a closed writer")
-	}
-	return statusToError(C.DBSstFileWriterOpen(fw.fw, goToCSlice([]byte(path))))
-}
-
-// Add puts a kv entry into the sstable being built. An error is returned if it
-// is not greater than any previously added entry (according to the comparator
-// configured during writer creation). `Open` must have been called. `Close`
-// cannot have been called.
-func (fw *RocksDBSstFileWriter) Add(kv MVCCKeyValue) error {
-	if fw == nil {
-		return errors.New("cannot call Open on a closed writer")
-	}
-	fw.DataSize += int64(len(kv.Key.Key)) + int64(len(kv.Value))
-	return statusToError(C.DBSstFileWriterAdd(fw.fw, goToCKey(kv.Key), goToCSlice(kv.Value)))
-}
-
-// Close finishes the writer, flushing any remaining writes to disk. At least
-// one kv entry must have been added.
-func (fw *RocksDBSstFileWriter) Close() error {
-	if fw.fw == nil {
-		return errors.New("writer is already closed")
-	}
-	err := statusToError(C.DBSstFileWriterClose(fw.fw))
-	fw.fw = nil
-	return err
 }
