+++ pkg/storage/engine/rocksdb/db_cpp.h
@@ -1,0 +1,28 @@
+// Copyright 2014 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied.  See the License for the specific language governing
+// permissions and limitations under the License.
+//
+// Author: Peter Mattis (peter@cockroachlabs.com)
+
+#include "db.h"
+#include "rocksdb/db.h"
+#include "rocksdb/comparator.h"
+
+::rocksdb::DB* RawDB(DBEngine *engine);
+::rocksdb::Comparator* CockroachComparator();
+
+std::string ToString(DBSlice s);
+rocksdb::Slice ToSlice(DBSlice s);
+rocksdb::Slice ToSlice(DBString s);
+DBStatus ToDBStatus(const rocksdb::Status& status);
+std::string EncodeKey(DBKey k);
